pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "ledmatrix"
version = "1.1.0"
description = "LED matrix display controller for Raspberry Pi"
requires-python = ">=3.10"
dependencies = [
    # Image processing
    "Pillow>=10.4.0,<12.0.0",
    "numpy>=1.24.0",

    # Timezone handling
    "pytz>=2024.2,<2025.0",
    "timezonefinder>=6.5.0,<7.0.0",
    "geopy>=2.4.1,<3.0.0",

    # HTTP requests
    "requests>=2.32.0,<3.0.0",

    # Google API integration
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "google-api-python-client>=2.147.0,<3.0.0",

    # Font rendering
    "freetype-py>=2.5.1,<3.0.0",

    # Spotify integration
    "spotipy>=2.24.0,<3.0.0",

    # Web framework
    "Flask>=3.0.0,<4.0.0",
    "werkzeug>=3.0.0,<4.0.0",
    "flask-wtf>=1.2.0",
    "flask-limiter>=3.5.0",

    # Text processing
    "unidecode>=1.3.8,<2.0.0",

    # Calendar integration
    "icalevents>=0.1.27,<1.0.0",

    # WebSocket support
    "python-socketio>=5.11.0,<6.0.0",
    "python-engineio>=4.9.0,<5.0.0",
    "websockets>=12.0,<14.0",
    "websocket-client>=1.8.0,<2.0.0",

    # JSON Schema validation
    "jsonschema>=4.20.0,<5.0.0",

    # System monitoring (web interface)
    "psutil>=6.0.0,<7.0.0",

    # Date/time utilities (web interface)
    "python-dateutil>=2.9.0,<3.0.0",
]

[project.optional-dependencies]
emulator = [
    "RGBMatrixEmulator",
]
hardware = [
    # rpi-rgb-led-matrix can be pip-installed from source:
    #   pip install git+https://github.com/hzeller/rpi-rgb-led-matrix
    # Requires C build tools on Raspberry Pi. Not added as a direct dependency
    # because it would fail resolution on non-ARM platforms. Install manually
    # on Pi hardware, or use: uv pip install git+https://github.com/hzeller/rpi-rgb-led-matrix
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.5.0",
    "pre-commit>=3.7.0",
    "types-requests",
    "types-pytz",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pip-audit==2.7.3",
]
test = [
    "pytest>=7.4.0,<9",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "click>=8.0.0",
]

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["test"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=30",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, may require external services)",
    "hardware: Tests that require actual hardware",
    "slow: Tests that take a long time to run",
    "plugin: Plugin-related tests",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.10"
platform = "linux"
follow_imports = "normal"
ignore_missing_imports = true
show_error_codes = true
warn_unused_ignores = true
warn_unreachable = true
strict_optional = true
disallow_untyped_defs = false
disallow_untyped_calls = false
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_configs = true
exclude = "(?x)(^test/|^.*/__pycache__/|^.*\\.pyc$)"

[[tool.mypy.overrides]]
module = "src.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "src.cache_manager"
disallow_untyped_defs = true
disallow_untyped_calls = true

[[tool.mypy.overrides]]
module = "src.config_manager"
disallow_untyped_defs = true
disallow_untyped_calls = true

[[tool.mypy.overrides]]
module = "src.plugin_system.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "src.exceptions"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "rgbmatrix.*",
    "RGBMatrixEmulator.*",
    "PIL.*",
    "freetype.*",
    "pytz.*",
    "requests.*",
    "Flask.*",
    "google.*",
    "spotipy.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]
ignore = [
    "E501",   # line too long (handled by formatter)
    # Pre-existing issues — TODO: fix in dedicated cleanup ticket (SPIKE-006)
    "B007",   # unused loop control variable
    "B023",   # binding loop variable in closure
    "B027",   # empty method without abstract decorator
    "B905",   # zip() without strict parameter
    "E402",   # module-level import not at top of file
    "E722",   # bare except
    "F401",   # unused import
    "F821",   # undefined name
    "F841",   # unused variable
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
